schema-snapshot: extract export-data exit hook into a named function
Move the inline atexit.Register closure out of exportData() into
registerExportDataExitSnapshotHook() in schemaSnapshotCapture.go, so the
call site is a single call and the main export function isn't polluted by
the exit-capture logic. Pure move — no behavior change.

Addresses review feedback on exportData.go.
